requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
transformers>=4.40.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pillow>=10.0.0
pdfplumber>=0.10.3
PyMuPDF>=1.22.5
torch>=2.0.0
